Now file will only write if CSV doesn't exist yet.

+++ First_Eclipse/src/firsteclipse/GenerateCSV.java
@@ -1,5 +1,6 @@
 package firsteclipse;
 
+import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
 
@@ -7,24 +8,31 @@ import sun.net.www.content.text.plain;
 
 public class GenerateCSV {
 	public static void save(String fileName) {
-		generateCsvFile("C:/Users/Me/Google Drive/UCL/III_Semester/Final Thesis/DOCUMENTATION/GeneratedData/" + fileName + ".csv");
+		generateCsvFile("C:/Users/Me/Google Drive/UCL/III_Semester/Final Thesis/DOCUMENTATION/GeneratedData/" + fileName
+				+ ".csv");
 	}
 
 	private static void generateCsvFile(String sFileName) {
-		try {
-			FileWriter writer = new FileWriter(sFileName);
+		File f = new File(sFileName);
+		if (!f.isFile()) { // If the file doesn't exist yet
 
-			for (String s : Glv.toNN) {
-				writer.append(s);
+			try {
+				FileWriter writer = new FileWriter(sFileName);
+
+				for (String s : Glv.toNN) {
+					writer.append(s);
+				}
+				// writer.append("DisplayName");
+
+				writer.flush();
+				writer.close();
+				System.out.println("Done writing CSV.");
+
+			} catch (IOException e) {
+				e.printStackTrace();
 			}
-			// writer.append("DisplayName");
+		} else
+			System.out.println("File already exists.");
 
-			writer.flush();
-			writer.close();
-			System.out.println("Done writing CSV.");
-			
-		} catch (IOException e) {
-			e.printStackTrace();
-		}
 	}
 }
